+++ packages/unlock-js/src/PublicLock/v4/purchaseKey.js
@@ -47,7 +47,7 @@ export default async function (
       erc20Address,
       lockAddress,
       this.provider,
-      this.signer
+      this.signer.address
     )
     if (!approvedAmount || approvedAmount.lt(actualAmount)) {
       // We must wait for the transaction to pass if we want the next one to succeed!