+++ packages/unlock-js/src/PublicLock/v6/purchaseKey.js
@@ -59,7 +59,7 @@ export default async function (
       erc20Address,
       lockAddress,
       this.provider,
-      this.signer
+      this.signer.address
     )
     if (!approvedAmount || approvedAmount.lt(actualAmount)) {
       // We must wait for the transaction to pass if we want the next one to succeed!