+++ packages/unlock-js/src/PublicLock/v9/purchaseKey.js
@@ -67,7 +67,7 @@ export default async function (
       erc20Address,
       lockAddress,
       this.provider,
-      this.signer
+      this.signer.address
     )
     if (!approvedAmount || approvedAmount.lt(actualAmount)) {
       await (